requirements.txt
# requirements.txt
streamlit>=1.38.0
deep-translator==1.11.4
requests==2.32.3
gtts==2.5.3
groq==0.4.1
python-dotenv==1.0.1
httpx==0.27.0
folium==0.15.1
streamlit-folium==0.17.0
speechrecognition
langdetect
tensorflow>=2.11.0
numpy>=1.23.0
pillow>=10.0.0
streamlit-modal
streamlit-float
Pyrebase4